requirements.txt
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.33
faiss-cpu==1.8.0
pypdf==4.2.0
python-dotenv==1.0.1
tiktoken==0.6.0
